requirements.txt
Flask
Flask-SocketIO
pymongo
python-dotenv
gunicorn
eventlet